UniformGrid: change algorithm to compute hash
This is a cute bug.

Before, the has was computed with a bitwise XOR operation. This is mostly fine.
The only problem with this is that XOR is commutative: a^b = b^a, so two
different objects could produce the same hash.

Suppose we only have x0 and x1. With XOR, x0^x1 = x1^x0, which is not good. Now,
we use Python's tuple hashing method to preserve the meaning of the various
variables.