Fix: allow more general names for GravitationalWaves
Before this commit, GravitationalWaves were found by identically matching the
Psi4 name in the outputs of the Multipole thorn. Now, this is more general and
we only require for the word psi4 to appear in the name. For instance, NP_Psi4
will match.

Thanks to Cheng-Hsin Cheng for pointing out this.